creating an arch bootstrapper
Creating a script to bootstrap arch installation.
Trying to balance security and usability.
Currently working on:
        - FDE
        - safe(er) partition mounts
        - linux-hardened kernel

Need to put a pause on the script so far.
The base archiso doesn't have zfs packages and it doesnt have space to
        install those packages.

I'm debating on expanding the ramdisk or creating a new iso.